fix bad unwrap in graph_view for hover_pos (#175)
This caused a panic for me when switched to the graph view in my app. It
tried to load a lot of nodes at once (called amongst UI code). Might
have been related to clicking on the app while it was loading, then
moving my cursor off, and when the event was processed there was no
hover_pos. (It returns None if the cursor is outside the window)

Co-authored-by: starlord <blitzar90@gmail.com>